PP-240: Renamed file CreateOffer
